read/write_dma: do not pass vector to lambda for iov version
Since we can pass this lambda to another processors, it is not advisable that
we move a vector inside the lambda. The said vector will be destructed when the
lambda finishes, and at that point it can be in a different processor.

In theory, we can use foreign pointers to achieve this. However, this is less
than ideal because foreign pointers can only be moved, and not copied. And
submit_io_write, submit_io_read, and submit_io, all take that function by
value.  Although this could be changed, passing them by value is a sensible
decision.

Moreover and more importantly, there is no reason really to use a foreign
pointer: All we need to do is to keep the object alive ourselves. The lambda
only needs to know the contents of the pointer.

Signed-off-by: Glauber Costa <glauber@scylladb.com>