IO queues
So far we have tested seastar and its child applications in environments where
the disk was fast enough to consume however many requests the processors would
throw at it.

Once we have started running on machines with slow disks, the situation changes
considerably: the processor is fast and keeps throwing requests at the disk,
but since the disk won't process them in a timely fashion they just pile up.

This situation is made much worse by the fact that modern SSDs will increase
their throughput with more requests only up to a certain point. After that, by
throwing more requests at the disk we are merely trashing it. Latencies will
increase for all requests without any visible increase in throughput. That
makes the request piling up worse.

The solution for this problem will vary from application to application.
However, it will likely involve throttling incoming requests. Also, since we
want to avoid trashing the disk's performance, we have to queue the requests
ourselves instead of allowing them all to be queued at the disk. Once we do
that, we can easily export the information about queued requests to the
application above us. It should know how to act on it.

Despite the throttling being the application's business, in my proposal the
core infrastructure lives in seastar proper. Although it is entirely possible
to leave seastar alone in favor of an application-only request queue that will
just prevent requests to enter our IO infrastructure, I have favoured a
seastar-based solution because:

  1) it is common infrastructure that other applications would eventually need,
  2) seastar is in a better position to make sure all requests are accounted for. An
     application side solution would have to capture writes and reads in potentially
     many places increasing complexity.

To queue the requests, this patchset introduces an "IO queue". An IO queue is the unit
responsible for coordinating IO initiation on behalf of a group of processors. In
systems that feature fast enough disks, we expect that each processor will handle its
own IO and the number of IO queue will equal the number of processors. In slower disks,
the number of requests needed to saturate the disk can be very low. In such scenarios,
having one IO queue per processor would force them to serialize everything even if the
other processors are not issuing any requests. Therefore, we expect those to run with
a smaller number of IO queue.

Once we have request queueing, the logical next step is allowing for request
priority.  The current solution treats all requests the same, and does not even
differentiate between reads and writes.

Future work:
- separate requests in separate priority queues,
- dynamically discover the amount of IO queue and requests the disk can handle.

Signed-off-by: Glauber Costa <glauber@scylladb.com>