resources: move more of hwloc functions inside a proper C++ class
The way this code currently stands, it is quite hard to reuse it.  In
preparation to reuse more of it, abstract it further into a C++ class - which
will have the added benefit of handling the destructors better.

It is far from perfect and we should improve this further. But that's a good
step and needed for a fix.

Signed-off-by: Glauber Costa <glauber@scylladb.com>