io queues: don't try to place them in processors that are not running
We should be restricting the processors in which we run. Not doing so will
break in cases where we are using -cpuset to restrict the processors in which
seastar is running.

We have a handy method in the hardware_topology class, so all we need to do
is call it.

Signed-off-by: Glauber Costa <glauber@scylladb.com>