resources: reuse topology for allocating io queues
Now that we return both CPUs and IO queues together, we don't have to reinitialize
the topology. Not only this is simpler, changes made to the topology - like restricting
to a specific CPU set, will be conserved.

This fixes a bug that would cause the alloation of the IO queues to break when --cpuset
was used.

Signed-off-by: Glauber Costa <glauber@scylladb.com>