IO queues: estimate disk latency
Estimating disk latency is useful for a variety of reasons: we may use it in
the mid term future to dynamically tune our IO queue capacity, or an
application throttling work based on accumulation of disk IOs can use that
information to decide for how long to throttle: boxes with faster disks would
throttle less, while boxes with slower disks would throttle more.

Measuring latencies for every request is expensive, and that is even more so in
virtualized environments with unstable TSC clocks. In an attempt to mitigate
this, we are not measuring the latency of every request, but rather, sampling
them.

Once a request is measured, we can't just use this as a latency figures since
spikes are common. To smooth that out, we use a time-based exponential moving
average over the measurements.

Using ScyllaDB's perf_sstable test (an IO intensive microbenchmark), I see
a ~2 % drop in performance with this estimator. This can probably be made cheaper
by not estimating the exponential with a table.

Signed-off-by: Glauber Costa <glauber@scylladb.com>