Kalman filter estimation of the latency of a request
Estimating disk latency is useful for a variety of reasons: we may use it in
the mid term future to dynamically tune our IO queue capacity, or an
application throttling work based on accumulation of disk IOs can use that
information to decide for how long to throttle: boxes with faster disks would
throttle less, while boxes with slower disks would throttle more.

Measuring latencies for every request is expensive, and that is even more so in
virtualized environments with unstable TSC clocks. In an attempt to mitigate
this, we are not measuring the latency of every request, but rather, sampling
them. The way we will do this, is by sampling every time we call the aio flusher
from the poller.

Signed-off-by: Glauber Costa <glauber@scylladb.com>