fair queue scheduling with priority classes
This is a proposal for a fair queue class, mainly intended to be used by Seastar's
IO queues to provide some form of fair scheduling between multiple agents.

An example of where this is needed arises early, in scenarios in which the number
of IO queues is less than the number of processors. We would like to see all processors
being served equally, but in reality nothing guarantees that. I've seen in practice
some processors lagging behind not being able to put their requests through.

More interesting uses arise from application code. It is customary to have many
different generators for IO. Without any kind of scheduling it is hard to guarantee
serviceability to any of them, or even throttle execution.

This initial implementation will do just that: given a specific balancing period,
the resolution of the queue, the fair queue class will guarantee that incoming requests
from the various priority_class'es will be served in rates proportional to their power.

It is up to the user of this class to make sure that the powers add up to
sensible values that represent the desired distribution.

TODO:
1) allow for the power of a queue to be dynamically changed
2) Finish integration with the IO queues. I have some code for that, but I wanted to have
   feedback on this first

Signed-off-by: Glauber Costa <glauber@scylladb.com>