io queues: use a fair queue instead of a semaphore
We now have a fair queue class. We can use it to implement our IO queues.  The
way it is right now, each processor that shares the IO queue will be assigned
to a default priority class with power of 1. That default class is applied to
all IOs equally within the same processor. Therefore, all it does is to achieve
fairness among processors within the same IO queue.

It is up to applications interested in fair queueing to add their own priority
classes and manage them accordingly.

Signed-off-by: Glauber Costa <glauber@scylladb.com>