reactor: destroy the I/O queues explicitly
To prevent ordering issues from rising, destroy the I/O queue explicitly at the
end of the reactor run.  This is needed because the reactor is destroyed from
the thread_local destructors. If the I/O queue happens to use any other
infrastructure that is also kept this way (for instance, collectd), we will not
have any way to guarantee who is destroyed first.

Signed-off-by: Glauber Costa <glauber@scylladb.com>