estimate proper I/O queue parameters
Introduction
------------

This is my first attempt to tackle the problem of estimating what are the proper
parameters for running the Seastar I/O queue.

I am still not trusting 100 % the results I am getting with this, and I need
to at the very least run this on more machines so I can increase my confidence.

Algorithm
---------

The main problem of course is that for the same machine, running the estimator
many times will lead to (sometimes hugely) different results.

But the algorithm that goes in here leads to the best results among the ones
I have tried, in terms of reducing inter-run variance.

The idea is as follows:
- compute the latency of the requests as we go
- remove the outliers
- based on the total time spent, calculate the number of requests we have
  issued.
- find the point where the loss in latency outgrows the gain in IOPS.

This is, we hope, the point where the curve starts to bend. We want to
run the I/O queue further down, so we multiply that result by the
empirical magic number "3".

To speed up the run, we run only every two requests until we see our first decrease,
and then go one by one.

I am writing to a 1GB file, and then reading requests for 200ms. I have tried a 10GB
file, reading requests for 300, 500, and 1000ms and didn't see a significant different.
I therefore kept the smaller values.

Application
-----------

I am trying to use as little seastar infrastructure as possible here. The idea is
that this code can be either ran during Seastar's initialization (if a parameter
is passed), or totally offline in a specifically designed tuning app. At the moment
I am not advocating one or the other, but I am trying to cast code that would work
both ways.

Early Results
-------------

For EC2.c3.x8large, this code will peak at around 10 requests. That will then lead
to a configuration value of 30 - 33 requests, which is in line with our former
recommendations.

For EC2.i2.x8large, this code will peak at around 16 - 18 requests. That will then lead
to a configuration value of 48 - 54 requests. This is far less than what we have previously
recommended, and I am still trying to figure out whether this is too low, or if our previous
recommendation was simply too high.

For my server's SSD, a Samsung 850 PRO, this yields 12 requests, which will then lead
to a configuration value of 42 requests. This one seem to be the most stable of them all,
I haven't seen a different value, really, unless I change the algorithm.

For my laptop's SSD, the result seems to be a random number.

Signed-off-by: Glauber Costa <glauber@scylladb.com>