thread pool: assert that we have successfully pushed to the spsc queue
The boost documentation says about push's return value:

    true, if the pop operation is successful, false if ringbuffer was empty.

We always expect it to be the case. And if so, let's assert it. If for some
reason this turns out not to be true, the consequences can include an angry
developer losing multiple debugging days.

Seastar issue #124 describes a scenario in which it happens. One of these
asserts will trigger in the test case described in that issue.

Signed-off-by: Glauber Costa <glauber@scylladb.com>