iotune: set a timeout, and uphold it
We have been running for way too long when presented with suboptimal conditions,
like (way too) slow disks. A description of a scenario where this generated
a problem is ScyllaDB's io-setup script hanging. An issue that describes this
can be found in ScyllaDB's issue tracker. See scylladb/scylla#1047 for details.

By default, we will run for 6 minutes. The reason for the choice of time is
explained in the aforementioned issue.

Signed-off-by: Glauber Costa <glauber@scylladb.com>