seastar::async allow passing an attribute
The old signature is kept around, but a new signature is added that allow us
to specify a thread_attribute. Users who need a thread_attribute to control their
threads at the moment have to resort to maintaining the threads manually.

Using seastar::async is a lot more convenient, and we should do it.

Signed-off-by: Glauber Costa <glauber@scylladb.com>