reactor: add threads to a separate background task list
Threads tend to be processor-heavy, while normal tasks are very light and
latency sensitive. Although a thread can yield often, it could easily be
the case in a system where a handful of threads are cooperating that a couple
of them will be placed in the _pending_tasks queue. Even if each of them run
for just a little bit of time, all of them together can add up to quite a bit.

In this patch, we'll augment the reactor with a new background tasks list. Tasks
that sit in the background task list will be run after the foreground tasks, with
a poller invocation after each of them, guaranteeing that latency sensitive tasks
will not be starved by backround tasks.

While in theory background and foreground tasks could be of any kind, it is my view
that we should put threads as background tasks, leaving the others as foreground.

Signed-off-by: Glauber Costa <glauber@scylladb.com>