syscall thread: do not use signals: pool instead
The syscall thread as it is today communicates with the main thread by means of
signal. That is, however, not necessary.

No signal is needed to wake up the syscall thread, since writing to the thread
pool's eventfd takes care of waking it up. We already do that, and this
direction of communication is fine.

For the other direction, however, we are currently using a signal to the
reactor. There is, however, no need to interrupt the reactor: we can just poll
like we do for all other sorts of interrupt-like events, and if there are
completed tasks, we'll queue them. We just need to make sure we resort back to
interrupt mode (signals) if the reactor goes to sleep.

One possible objection to this is that the reactor may never pool, and we may
take a long time to get completions. While that is true, this is also a generic
problem shared by all pollers, and one that we solve by means of the task quota.

Writing the syscall thread like that yields improvements of up to 40 % to in a
simple benchmark that computes the cost of calling file_size() in a loop.

Signed-off-by: Glauber Costa <glauber@scylladb.com>