don't always return true for should_yield on seastar threads
Currently, we have two main kinds of threads - as defined by their attributes:
threads that belong to a scheduling group, and threads that don't. For the
latter, should_yield() has a very precise meaning: if we are still within the
execution time determined by the schedule group, then we should yield.

For threads that do not belong to a scheduling group, this meaning is less
solid and we currently just return true all the time.

I will argue, though, that we can do better than that: seastar constantly
keeps a tab on whether or not we need to preempt the current task out of the
processor. The future core infrastructure use that to determine whether or
not we should break continuations or keep executing them.

We should do the same for threads. This patch changes the should_yield()
method of threads not currently beloning to a scheduling group so that
it only returns true when we have effectively ran for too long.

Signed-off-by: Glauber Costa <glauber@scylladb.com>