threads: move thread impl to its own file
The thread_impl namespace exists so that the core infrastructure can call in to
some of the thread infrastructure without having knowledge of the full thing -
which would certainly generate nasty circular include dependencies.

Right now, those signatures are copied into future.hh which means that if we
need to add or remove some, we need to do this in more than one place. But even
more important, this guarantees that every time we need to use them we will
generate an explicit out of line call with no chance of inlining.

Some of those operations are simple enough and frequent enough that we would
like to, if possible, do some inlining.

Signed-off-by: Glauber Costa <glauber@scylladb.com>