percentile accumulator: a helper class to allow easy tracking of latency percentiles
Some of our collectd metrics export latency metrics: the time it takes for an
operation to complete, the time requests spend sitting in a queue, etc. The
problem with that approach is that collectd is by nature a sampling mechanism,
and only the value that happens to be stored when we poll will be read.

This is not so much of an issue if we have control of the polling period, as we
can try to make it quite small when debugging issues. However, with very large
poll periods those counters can be quite misleading, as they will smooth out the
measurements and potentially hide peaks that are responsible for things like
timeouts, etc.

The solution for that is to start reporting percentiles. Collectd per se can
only export scalar values, but we can easily add multiple counters per metric,
one per each percentile.

Another requirement is that we don't store each and every value that we see, as
the memory requirements for that can easily grow fast.

This patch uses boost accumulator's implementation of the p_square algorithm to
approximate distribution quantiles. We wrap that in a format that is more
convenient for the uses cases we have in mind, specially for exporting those
metrics over collectd.

An example of collectd registration is as follows (from ScyllaDB, out of tree):

    struct dirty_memory_manager {
        [ ... ]
        using latency_percentile_accumulator_type = seastar::percentile_accumulator<50, 90, 95, 99, 100>;
        latency_percentile_accumulator_type _blocked_for;

        [ ... ]
        latency_percentile_accumulator& percentile_accumulator() { return _blocked_for; }
        [ ... ]
        void add_measurement(double m) { _blocked_for.add_measurement(m); }
    };

    _dirty_memory_manager.percentile_accumulator().for_each_accumulator([this] (auto& acc) {
        _collectd.push_back(
            scollectd::add_polled_metric(scollectd::type_instance_id("database"
                    , scollectd::per_cpu_plugin_instance
                    , "delay", "requests_blocked_memory_" + acc.percentile_sstring())
                    , scollectd::make_typed(scollectd::data_type::GAUGE, [this, &acc] {
                        return acc.probe_for_period();
                    })
        ));
    });

The code above will generate the following metrics:

  - database-*/delay-requests_blocked_memory_50
  - database-*/delay-requests_blocked_memory_90
  - database-*/delay-requests_blocked_memory_95
  - database-*/delay-requests_blocked_memory_99
  - database-*/delay-requests_blocked_memory_max

And the user tracking this latency needs only to call add_measurement() once.

Signed-off-by: Glauber Costa <glauber@scylladb.com>