iotune: do not spawn worker thread for first worker
For iotune to work, we need to have n independent workers pushing
concurrency into the disk. We do that by spawning one thread per
available core and assigning work to them.

The problem with that is that those threads can throw exceptions, and
the way exceptions work, they will not be propagated to the main thread.
That is not a problem for most exceptions: they indicate that something
has gone wrong and terminating the whole program is the only sane course
of action.

But for other exceptions, like the timeout exception, that is not the
case. Exceptions that are expected - like the timeout - are only ever
thrown by the thread with id == 0, since they are explicitly thrown by
the control logic.

Instead of complicating our flow by storing the exception somewhere and
then tracking it down in the main thread, it is easier to just make sure
that the coordinator is not running in a separate thread. This is a
common pattern and the seastar reactor is one example of where such
pattern is employed.

Signed-off-by: Glauber Costa <glauber@scylladb.com>