iotune: terminate gracefully before throwing an exception
If we throw a timeout exception, we will catch it and attempt to return
an error code. But because the other threads will still be running at
this point, this attempt will fail and we'll throw when noticing that
the threads were not joined.

Detaching them doesn't really help since the barrier destructors won't
like that very much - if it so happened that one of the threads made
into one of the barriers.

So in this patch we just terminate gracefully before we throw.

Signed-off-by: Glauber Costa <glauber@scylladb.com>