queue: allow queue to grow and be forceably destroyed
This patch adds two new methods for the seastar queue: expand(), will
increase the size of the queue. This is useful in situations in which
we have queues of objects that are expensive to maintain, so we'd want
to keep the queue small until we are positive that we really need to
grow it.

We also add a clear() method, that forces the current elements in the
queue to be dropped. Useful during shutdown conditions.

Signed-off-by: Glauber Costa <glauber@scylladb.com>