iotune: make sure help is working
bpo::notify() will throw if we just call iotune with:

   iotune --help

This is because the evaluation directory is mandatory, and it won't be
present in the above command line. We end up with a confusing message,
telling us to try --help, which is what we have just did.

Move it earlier, so it works.

Signed-off-by: Glauber Costa <glauber@scylladb.com>