fair_queue: allow execute on empty queues
Right now, execute will always be called exactly as many times as
we have requests in the queue.

It's better not to rely on that. We could, for instance, start thinking
about taking requests that are taking too long out of the queue -
something that we have been using as an example of the I/O queue
capabilities but haven't done yet.

But in a more immediate goal, we'll start acking request termination
explicitly, and when the semaphore is granted it may not have any more
requests in the queue.

Signed-off-by: Glauber Costa <glauber@scylladb.com>