issue new I/Os as soon as we complete the previous ones
Currently, when an I/O operation complete we will make its future ready,
which will in turn schedule another continuation that will finally
release the fair queue semaphore - allowing yet another continuation to
be scheduled for the waiter who will submit another I/O operation.

This is particularly bad when the server is overloaded. In those
situations, the new continuation could sit behind a storm of existing
continuations and it could be a while until we get to it. During that
time, no new I/Os will be issued. We need to issue as soon as we can.

One thing to notice is that when the fair queue selects a new request,
we are not actually submitting anything: the submission will come in the
next continuation. To handle that, we'll try to induce that continuation
to be scheduled as urgent. Since the methods to do so are private, we'll
employ a trick to forward from an available future. Alternatively, we
could store the function and execute right away. I just can't see a way
to do it without an allocation in the submit path.

I have tested this with the soon-to-be merged iotest, simulating a load
in which write requests will fight cpu-bound tasks for the reactor. This
patch improves the IOPS, and consequentially throughput twofold.

ScyllaDB saw around 5 % improvement with a workload that generates
around 1k pending tasks in the reactor.

Signed-off-by: Glauber Costa <glauber@scylladb.com>