allow seastar application to override parameter's default
By passing new value_semantic pointers, this patch allows any seastar
application to change defaults for existing options.

This is done by passing an unordered map with defaults. Seastar, upon
finding an option with the said name, will replace the original
value_semantic with a new one.

Technically, passing the whole value_semantic is overly flexible: all we
need is to pass the value. However, it is quite a challenge to derive the
type of an option from boost program_options. So we will go this route
instead.

Signed-off-by: Glauber Costa <glauber@scylladb.com>