allow applications to override task-quota-ms
The default task quota is quite conservative, and may not be the best
for all applications. Some, specially the more latency-sensitive ones,
will benefit from lower defaults.

This patch adds the ability for an app to ask seastar for a lower
default, which will still can be overridden by the user.

Example usage:

    app_template::config cfg;
    cfg.default_task_quota = 500us;
    app_template app(std::move(cfg));

Signed-off-by: Glauber Costa <glauber@scylladb.com>