io_queue: move away from semaphore.
This patch does away with the semaphore in the fair_queue / IO queue and
substitute a request/dispatch mechanism for that. In itself, it should
allow for better latencies as this reduces the amount of continuations
in the I/O path and allow us to dispatch I/O from the seastar poller.

But the way it is really used is as stepping stone towards changing the
I/O queue to accept rate limits - while keeping compatibility with what
we already have. We will do that by adding additional constraints to
the I/O dispatcher. That is hard to do with the semaphore-based
approach, but easy to do with the proposed approach.

Signed-off-by: Glauber Costa <glauber@scylladb.com>