util: implementation of the token_bucket algorithm
The token bucket algorithm can be used to rate-limit a bursty process
given a desired throuhgput and the maximum period in which to handle
bursts.

As we move the I/O queues more towards limiting bandwidth and IOPS
instead of enforcing concurrent requests, rate limiting will play a
central role in the I/O subsystem.

Signed-off-by: Glauber Costa <glauber@scylladb.com>